+++ app/src/main/java/com/example/ihonboarding/feature/home/data_source/local/db/NewsDb.kt
@@ -1,8 +1,8 @@
-package com.example.ihonboarding.data_source.home.local.db
+package com.example.ihonboarding.feature.home.data_source.local.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.ihonboarding.data_source.home.local.dao.NewsDao
+import com.example.ihonboarding.feature.home.data_source.local.dao.NewsDao
 import com.example.ihonboarding.domain.home.model.News
 
 @Database(entities = [News::class], version = 1)
