+++ app/src/main/java/com/example/ihonboarding/feature/home/data_source/remote/api/NewsService.kt
@@ -1,6 +1,6 @@
-package com.example.ihonboarding.data_source.home.remote.api
+package com.example.ihonboarding.feature.home.data_source.remote.api
 
-import com.example.ihonboarding.data_source.home.remote.dto.NewsListDto
+import com.example.ihonboarding.feature.home.data_source.remote.dto.NewsListDto
 import retrofit2.http.GET
 import retrofit2.http.Header
