+++ app/src/main/java/com/example/ihonboarding/feature/home/data_source/remote/dto/CounterDto.kt
@@ -1,4 +1,4 @@
-package com.example.ihonboarding.data_source.home.remote.dto
+package com.example.ihonboarding.feature.home.data_source.remote.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
